% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDEE2.R
\name{queryDee2}
\alias{queryDee2}
\title{Query Whether a DEE2 Dataset is Available}
\usage{
queryDee2(species, SRRvec, metadata = NULL, ...)
}
\arguments{
\item{x}{A character string matching a species of interest.}

\item{y}{A character string or vector thereof of SRA run accession numbers.}
}
\description{
This function sends a query to check whether a dataset is available or not.
}
\examples{
x<-queryDee2("ecoli",c("SRR1067773","SRR5350513"))
}
\keyword{query}
